LPS-196226 Transform JSBundleConfigRegistry into JSBundleConfigRegistryUtil, change JSBundleConfigServiceTrackerCustomizer interface to be EagerServiceTrackerCustomizer so the tracker is open immediately
